rename to IDigitalManifestation, and fix namespaces
