+++ src/Wellcome.Dds/DlcsJobProcessor/DashboardContinuousRunningStrategy.cs
@@ -5,7 +5,7 @@ using Microsoft.Extensions.Options;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.Extensions.DependencyInjection;
-using Wellcome.Dds.AssetDomain.Dashboard;
+using Wellcome.Dds.AssetDomain.DigitalObjects;
 using Wellcome.Dds.AssetDomain.Dlcs.Ingest;
 
 namespace DlcsJobProcessor
