+++ src/Wellcome.Dds/Wellcome.Dds.AssetDomain/DigitalObjects/IDigitalObject.cs
@@ -1,4 +1,4 @@
-﻿namespace Wellcome.Dds.AssetDomain.Dashboard
+﻿namespace Wellcome.Dds.AssetDomain.DigitalObjects
 {
     public interface IDigitalObject
     {
