+++ src/Wellcome.Dds/Wellcome.Dds.AssetDomain/DigitalObjects/IDigitisedCollection.cs
@@ -1,12 +1,12 @@
 ﻿using System.Collections.Generic;
 using Wellcome.Dds.AssetDomain.Mets;
 
-namespace Wellcome.Dds.AssetDomain.Dashboard
+namespace Wellcome.Dds.AssetDomain.DigitalObjects
 {
     public interface IDigitisedCollection : IDigitalObject
     {
         ICollection MetsCollection { get; set; }
-        IEnumerable<IDigitisedManifestation> Manifestations { get; set; }
+        IEnumerable<IDigitalManifestation> Manifestations { get; set; }
         IEnumerable<IDigitisedCollection> Collections { get; set; }
     }
 }
