+++ src/Wellcome.Dds/Wellcome.Dds.AssetDomain/DigitalObjects/IStatusProvider.cs
@@ -2,7 +2,7 @@
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace Wellcome.Dds.AssetDomain.Dashboard
+namespace Wellcome.Dds.AssetDomain.DigitalObjects
 {
     public interface IStatusProvider
     {
