+++ src/Wellcome.Dds/Wellcome.Dds.AssetDomain/DigitalObjects/JobActivity.cs
@@ -2,7 +2,7 @@
 using Wellcome.Dds.AssetDomain.Dlcs.Ingest;
 using Wellcome.Dds.AssetDomain.Dlcs.Model;
 
-namespace Wellcome.Dds.AssetDomain.Dashboard
+namespace Wellcome.Dds.AssetDomain.DigitalObjects
 {
     public class JobActivity
     {
