+++ src/Wellcome.Dds/Wellcome.Dds.AssetDomainRepositories.Tests/Dashboard/DatabaseStatusProviderTests.cs
@@ -7,7 +7,7 @@ using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging.Abstractions;
 using Microsoft.Extensions.Options;
 using Wellcome.Dds.AssetDomainRepositories.Control;
-using Wellcome.Dds.AssetDomainRepositories.Dashboard;
+using Wellcome.Dds.AssetDomainRepositories.DigitalObjects;
 using Wellcome.Dds.Common;
 using Xunit;
