+++ src/Wellcome.Dds/Wellcome.Dds.Dashboard/Controllers/AccountController.cs
@@ -1,6 +1,6 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using Utils;
-using Wellcome.Dds.AssetDomain.Dashboard;
+using Wellcome.Dds.AssetDomain.DigitalObjects;
 
 namespace Wellcome.Dds.Dashboard.Controllers
 {
