+++ src/Wellcome.Dds/Wellcome.Dds.Dashboard/Controllers/DashController.cs
@@ -9,7 +9,7 @@ using Newtonsoft.Json;
 using Utils;
 using Utils.Logging;
 using Wellcome.Dds.AssetDomain;
-using Wellcome.Dds.AssetDomain.Dashboard;
+using Wellcome.Dds.AssetDomain.DigitalObjects;
 using Wellcome.Dds.AssetDomain.Dlcs;
 using Wellcome.Dds.AssetDomain.Dlcs.Ingest;
 using Wellcome.Dds.AssetDomain.Dlcs.Model;
