+++ src/Wellcome.Dds/Wellcome.Dds.Dashboard/Controllers/DashGlobalsActionFilter.cs
@@ -3,7 +3,7 @@ using Microsoft.AspNetCore.Mvc.Filters;
 using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
-using Wellcome.Dds.AssetDomain.Dashboard;
+using Wellcome.Dds.AssetDomain.DigitalObjects;
 
 namespace Wellcome.Dds.Dashboard.Controllers
 {
