+++ src/Wellcome.Dds/Wellcome.Dds.Dashboard/Views/Dash/Collection.cshtml
@@ -1,4 +1,4 @@
-﻿@model IDigitisedCollection
+﻿@model Wellcome.Dds.AssetDomain.DigitalObjects.IDigitisedCollection
 
 @{
     ViewBag.Title = "Dash: " + Model.Identifier;
