+++ src/Wellcome.Dds/Wellcome.Dds.Dashboard/Views/_ViewImports.cshtml
@@ -4,7 +4,6 @@
 @using Wellcome.Dds.AssetDomain.Dlcs.Ingest
 @using Wellcome.Dds.Common
 @using Wellcome.Dds.AssetDomain
-@using Wellcome.Dds.AssetDomain.Dashboard
 @using Wellcome.Dds.AssetDomain.Dlcs
 @using Wellcome.Dds.AssetDomain.Dlcs.Model
 @using Wellcome.Dds.AssetDomain.Mets
