+++ src/Wellcome.Dds/Wellcome.Dds.Server/Models/SimpleCollectionModel.cs
@@ -1,6 +1,6 @@
 using System.Collections.Generic;
 using Utils;
-using Wellcome.Dds.AssetDomain.Dashboard;
+using Wellcome.Dds.AssetDomain.DigitalObjects;
 
 namespace Wellcome.Dds.Server.Models
 {
